% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add}
\alias{add}
\title{Add}
\usage{
add(a, b)
}
\arguments{
\item{a}{the first number}

\item{b}{the second number}
}
\value{
The sum of two numbers
}
\description{
Adds two numbers
}
\examples{
add(2, 3)
}
