% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{substract}
\alias{substract}
\title{Substracts two numbers}
\usage{
substract(a, b)
}
\arguments{
\item{a}{the first number}

\item{b}{the number to substract}
}
\value{
The difference of two numbers (a-b)
}
\description{
Substracts two numbers
}
\examples{
substract(4, 2)
}
